requirements.txt
slack-bolt
python-dotenv
google-genai
gunicorn
